+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+
 - Add support for Arbitrum-Mainnet
+- Add optional arguments to allow setting amount for payable contract method invocations
 
 ## [0.3.0] - 2024-09-05